@@ -16,10 +16,8 @@ version = "SNAPSHOT." + buildNumber
 group = "com.eliotlash.mclib"
 def repsyUrl = project.repsyUrl
 project.archivesBaseName
-sourceCompatibility = targetCompatibility = '1.8' // Need this here so eclipse task generates correctly.
-compileJava {
-    sourceCompatibility = targetCompatibility = '1.8'
-}
+
+java.toolchain.languageVersion = JavaLanguageVersion.of(17)
 
 repositories {
     // Use jcenter for resolving dependencies.
