@@ -2,15 +2,15 @@ package com.eliotlash.mclib.utils;
 
 public class MathUtils {
 	public static int clamp(int x, int min, int max) {
-		return x < min ? min : (x > max ? max : x);
+		return Math.min(Math.max(x, min), max);
 	}
 
 	public static float clamp(float x, float min, float max) {
-		return x < min ? min : (x > max ? max : x);
+		return Math.min(Math.max(x, min), max);
 	}
 
 	public static double clamp(double x, double min, double max) {
-		return x < min ? min : (x > max ? max : x);
+		return Math.min(Math.max(x, min), max);
 	}
 
 	public static int cycler(int x, int min, int max) {
